Added more build options

+++ dockerpi-toolchain.sh
@@ -19,6 +19,18 @@ then
     exit 1
 fi
 
+if [ -z "$TOOLCHAIN_PATH" ]
+then
+    echo "$ROOTFS_PATH not set"
+    exit 1
+fi
+
+if [ -z "$TARGET_ARCH" ]
+then
+    echo "$TARGET_ARCH not set"
+    exit 1
+fi
+
 echo "Remove exists $TOOLCHAIN_PATH"
 
 rm -rf $TOOLCHAIN_PATH
@@ -29,7 +41,7 @@ then
     exit 1
 fi
 
-docker build -f Dockerfile.toolchain --build-arg TOOLCHAIN_PATH_ARG=${TOOLCHAIN_PATH} --network=host -t dockerpi/toolchain-${TARGET_ARCH} .
+docker build -f Dockerfile.toolchain --build-arg TOOLCHAIN_PATH_ARG=${TOOLCHAIN_PATH} --build-arg TOOLCHAIN_ARCH_ARG=${TARGET_ARCH} --network=host -t dockerpi/toolchain-${TARGET_ARCH} .
 docker run -v /opt:/opt/mount --rm dockerpi/toolchain-${TARGET_ARCH} bash -c "cp -r ${TOOLCHAIN_PATH} /opt/mount/"
 ${TOOLCHAIN_PATH}/bin/arm-linux-gnueabihf-cpp --version
